Another large batch of changes.
Completely removed the daily hours quota since it was causing some headache on the refund end and we don't make use of hourly quotas beyond per-reservation.

Completely redesigned the default template, super simplifying by reducing the used colors to two (green and gray), removing the sidebar and legend, moving the buttons above the table, and adding stylyzed extras. Also switched to jquery-ui smoothness instead of redmond.

Removed residual files and sections related to the non-ajax login process. This functionality should be fairly easy to re-implement, but in the interest in keeping this a simple kiosk-only implementation (the original goal) they won't be included in this app.

This might possibly be the first release version or at least a stable enough copy, though it needs more testing.